requirements.txt
argparse
flask
pandas
requests
pybinaryedge
python-Wappalyzer
python-dotenv
shodan
flask-login
flask-wtf
flask-migrate
flask-sqlalchemy
flask-login
werkzeug